filter by id_kpi in ArchivedKpis

+++ Quantis.WorkFlow.Services/DTOs/API/ARulesDTO.cs
@@ -9,7 +9,7 @@ namespace Quantis.WorkFlow.Services.DTOs.API
         public int id_kpi { get; set; }
         public string name_kpi { get; set; }
         public DateTime interval_kpi { get; set; }
-        public int value_kpi { get; set; }
+        public double value_kpi { get; set; }
         public int ticket_id { get; set; }
         public DateTime close_timestamp_ticket { get; set; }
         public bool archived { get; set; }
